src/main/webapp/app/entities/quilombo/route/quilombo-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IQuilombo, Quilombo } from '../quilombo.model';
import { QuilomboService } from '../service/quilombo.service';

import { QuilomboRoutingResolveService } from './quilombo-routing-resolve.service';

describe('Quilombo routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: QuilomboRoutingResolveService;
  let service: QuilomboService;
  let resultQuilombo: IQuilombo | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(QuilomboRoutingResolveService);
    service = TestBed.inject(QuilomboService);
    resultQuilombo = undefined;
  });

  describe('resolve', () => {
    it('should return IQuilombo returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultQuilombo = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultQuilombo).toEqual({ id: 123 });
    });

    it('should return new IQuilombo if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultQuilombo = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultQuilombo).toEqual(new Quilombo());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Quilombo })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultQuilombo = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultQuilombo).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
